DualMode Client Loader
To be able to control when the Client side starts loading in a background frame.
I wanted a seperate component to control the loading of the client side dlls.
Need to register this with the DI container
Update the js to not call loadClient automatically but make it available for jsInterop.
Add it to the AppModel and initialize after async to kick of the timer