Removed package references from the API
Not sure why they are in the API.  Should only be required in the server but maybe I am missing something.  Anyway had to be removed to fix.

The project references the ASP.NET Core shared framework, which is not supported by Blazor WebAssembly apps. Remove the framework reference if directly referenced, or the package refe...